re-organize numpyop and tensorop location
