+++ fastestimator/backend/feed_forward.py
@@ -14,6 +14,7 @@
 # ==============================================================================
 from typing import TypeVar, Union
 
+import numpy as np
 import tensorflow as tf
 import torch
 
@@ -22,7 +23,8 @@ from fastestimator.backend.to_tensor import to_tensor
 Tensor = TypeVar('Tensor', tf.Tensor, torch.Tensor)
 
 
-def feed_forward(model: Union[tf.keras.Model, torch.nn.Module], x: Tensor, training: bool = True) -> Tensor:
+def feed_forward(model: Union[tf.keras.Model, torch.nn.Module], x: Union[Tensor, np.ndarray],
+                 training: bool = True) -> Tensor:
     if isinstance(model, tf.keras.Model):
         if not isinstance(x, tf.Tensor):
             x = to_tensor(x, "tf")
