+++ fastestimator/op/numpyop/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-from fastestimator.op.op import Op, get_current_ops, get_inputs_by_op, write_outputs_by_op
+from fastestimator.op.numpyop.numpyop import Delete, NumpyOp, forward_numpyop
