+++ fastestimator/op/numpyop/meta/one_of.py
@@ -17,7 +17,7 @@ from typing import Any, Dict, List, Union
 
 import numpy as np
 
-from fastestimator.op.op import NumpyOp
+from fastestimator.op.numpyop.numpyop import NumpyOp
 
 
 class OneOf(NumpyOp):
