+++ fastestimator/op/numpyop/meta/sometimes.py
@@ -16,7 +16,7 @@ from typing import Any, Dict, List, Union
 
 import numpy as np
 
-from fastestimator.op.op import NumpyOp
+from fastestimator.op.numpyop.numpyop import NumpyOp
 
 
 class Sometimes(NumpyOp):
