+++ fastestimator/op/numpyop/multivariate/multivariate.py
@@ -13,33 +13,12 @@
 # limitations under the License.
 # ==============================================================================
 from collections import OrderedDict
-from copy import deepcopy
 from typing import Any, Dict, List, Optional, Union
 
 import numpy as np
 
-from albumentations import BboxParams, Compose, DualTransform, ImageOnlyTransform, KeypointParams, ReplayCompose
-from fastestimator.op.op import NumpyOp
-
-
-class ImageOnlyAlbumentation(NumpyOp):
-    def __init__(self,
-                 func: ImageOnlyTransform,
-                 inputs: Union[List[str], str, None] = None,
-                 outputs: Union[List[str], str, None] = None,
-                 mode: Optional[str] = None):
-        super().__init__(inputs=inputs, outputs=outputs, mode=mode)
-        if isinstance(self.inputs, List) and isinstance(self.outputs, List):
-            assert len(self.inputs) == len(self.outputs), "Input and Output lengths must match"
-        self.func = Compose(transforms=[func])
-        self.replay_func = ReplayCompose(transforms=[deepcopy(func)])
-        self.in_list, self.out_list = True, True
-
-    def forward(self, data: List[np.ndarray], state: Dict[str, Any]) -> List[np.ndarray]:
-        results = [self.replay_func(image=data[0]) if len(data) > 1 else self.func(image=data[0])]
-        for i in range(1, len(data)):
-            results.append(self.replay_func.replay(results[0]['replay'], image=data[i]))
-        return [result["image"] for result in results]
+from albumentations import BboxParams, Compose, DualTransform, KeypointParams
+from fastestimator.op.numpyop.numpyop import NumpyOp
 
 
 class MultiVariateAlbumentation(NumpyOp):
