+++ fastestimator/op/numpyop/univariate/channel_dropout.py
@@ -16,7 +16,7 @@ from typing import Union, Iterable, Callable, Tuple
 
 from albumentations.augmentations.transforms import ChannelDropout as ChannelDropoutAlb
 
-from fastestimator.op.numpyop.base_augmentations import ImageOnlyAlbumentation
+from fastestimator.op.numpyop.univariate.univariate import ImageOnlyAlbumentation
 
 
 class ChannelDropout(ImageOnlyAlbumentation):
