+++ fastestimator/op/numpyop/univariate/from_float.py
@@ -16,7 +16,8 @@ from typing import Union, Iterable, Callable, Optional
 
 from albumentations.augmentations.transforms import FromFloat as FromFloatAlb
 
-from fastestimator.op.numpyop.base_augmentations import ImageOnlyAlbumentation
+from fastestimator.op.numpyop.univariate.univariate import ImageOnlyAlbumentation
+
 import numpy as np
 
 
@@ -27,7 +28,7 @@ class FromFloat(ImageOnlyAlbumentation):
             inputs: Key(s) of images to be normalized
             outputs: Key(s) of images to be normalized
             mode: What execution mode (train, eval, None) to apply this operation
-            max_value: The maximum value to serve as the divisor. If None it will be inferred by dtype. 
+            max_value: The maximum value to serve as the divisor. If None it will be inferred by dtype.
             dtype: the data type to cast the output
         Image types:
             float32
