+++ fastestimator/op/numpyop/univariate/hue_saturation_value.py
@@ -16,7 +16,7 @@ from typing import Union, Iterable, Callable, Tuple
 
 from albumentations.augmentations.transforms import HueSaturationValue as HueSaturationValueAlb
 
-from fastestimator.op.numpyop.base_augmentations import ImageOnlyAlbumentation
+from fastestimator.op.numpyop.univariate.univariate import ImageOnlyAlbumentation
 
 
 class HueSaturationValue(ImageOnlyAlbumentation):
