+++ fastestimator/op/numpyop/univariate/image_compression.py
@@ -16,7 +16,7 @@ from typing import Union, Iterable, Callable
 
 from albumentations.augmentations.transforms import ImageCompression as ImgCmpAlb
 
-from fastestimator.op.numpyop.base_augmentations import ImageOnlyAlbumentation
+from fastestimator.op.numpyop.univariate.univariate import ImageOnlyAlbumentation
 
 
 class ImageCompression(ImageOnlyAlbumentation):
