+++ fastestimator/op/numpyop/univariate/motion_blur.py
@@ -16,7 +16,7 @@ from typing import Union, Iterable, Callable, Tuple
 
 from albumentations.augmentations.transforms import MotionBlur as MotionBlurAlb
 
-from fastestimator.op.numpyop.base_augmentations import ImageOnlyAlbumentation
+from fastestimator.op.numpyop.univariate.univariate import ImageOnlyAlbumentation
 
 
 class MotionBlur(ImageOnlyAlbumentation):
