+++ fastestimator/op/numpyop/univariate/random_snow.py
@@ -16,7 +16,7 @@ from typing import Union, Iterable, Callable
 
 from albumentations.augmentations.transforms import RandomSnow as RandomSnowAlb
 
-from fastestimator.op.numpyop.base_augmentations import ImageOnlyAlbumentation
+from fastestimator.op.numpyop.univariate.univariate import ImageOnlyAlbumentation
 
 
 class RandomSnow(ImageOnlyAlbumentation):
