+++ fastestimator/op/numpyop/univariate/reshape.py
@@ -16,7 +16,7 @@ from typing import Any, Callable, Dict, Iterable, List, Tuple, Union
 
 import numpy as np
 
-from fastestimator.op.op import NumpyOp
+from fastestimator.op.numpyop.numpyop import NumpyOp
 
 
 class Reshape(NumpyOp):
