+++ fastestimator/op/numpyop/univariate/univariate.py
@@ -12,30 +12,30 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-from typing import Any, Callable, Dict, Iterable, List, Union
+from copy import deepcopy
+from typing import Any, Dict, List, Optional, Union
 
 import numpy as np
 
+from albumentations import Compose, ImageOnlyTransform, ReplayCompose
 from fastestimator.op.numpyop.numpyop import NumpyOp
 
 
-class ChannelTranspose(NumpyOp):
-    """Transpose the data (for example to make it channel-width-height instead of width-height-channel)
-
-    Args:
-            inputs: Key(s) of arrays to be transposed
-            outputs: Key(s) of arrays to be transposed
-            mode: What execution mode (train, eval, None) to apply this operation
-            axes: The permutation axes
-    """
+class ImageOnlyAlbumentation(NumpyOp):
     def __init__(self,
-                 inputs: Union[None, str, Iterable[str], Callable] = None,
-                 outputs: Union[None, str, Iterable[str]] = None,
-                 mode: Union[None, str, Iterable[str]] = None,
-                 axes: List[int] = (2, 0, 1)):
+                 func: ImageOnlyTransform,
+                 inputs: Union[List[str], str, None] = None,
+                 outputs: Union[List[str], str, None] = None,
+                 mode: Optional[str] = None):
         super().__init__(inputs=inputs, outputs=outputs, mode=mode)
-        self.axes = axes
+        if isinstance(self.inputs, List) and isinstance(self.outputs, List):
+            assert len(self.inputs) == len(self.outputs), "Input and Output lengths must match"
+        self.func = Compose(transforms=[func])
+        self.replay_func = ReplayCompose(transforms=[deepcopy(func)])
         self.in_list, self.out_list = True, True
 
     def forward(self, data: List[np.ndarray], state: Dict[str, Any]) -> List[np.ndarray]:
-        return [np.transpose(elem, self.axes) for elem in data]
+        results = [self.replay_func(image=data[0]) if len(data) > 1 else self.func(image=data[0])]
+        for i in range(1, len(data)):
+            results.append(self.replay_func.replay(results[0]['replay'], image=data[i]))
+        return [result["image"] for result in results]
