+++ fastestimator/op/tensorop/loss/cross_entropy.py
@@ -20,7 +20,7 @@ import torch
 from fastestimator.backend.binary_crossentropy import binary_crossentropy
 from fastestimator.backend.categorical_crossentropy import categorical_crossentropy
 from fastestimator.backend.sparse_categorical_crossentropy import sparse_categorical_crossentropy
-from fastestimator.op.op import TensorOp
+from fastestimator.op.tensorop.tensorop import TensorOp
 
 Tensor = TypeVar('Tensor', tf.Tensor, torch.Tensor)
