+++ fastestimator/op/tensorop/model/update.py
@@ -19,7 +19,7 @@ import torch
 from tensorflow.python.framework import ops as tfops
 
 from fastestimator.backend.update_model import update_model
-from fastestimator.op.op import TensorOp
+from fastestimator.op.tensorop.tensorop import TensorOp
 
 Tensor = TypeVar('Tensor', tf.Tensor, torch.Tensor)
